rename root source-code directory to src
