+++ src/utils/appUpdater.ts
@@ -1,7 +1,7 @@
 import { BrowserWindow, dialog, Tray } from 'electron';
 import { autoUpdater } from 'electron-updater';
 import log from 'electron-log';
-import { setIsAppQuitting, setIsUpdateAvailable } from '..';
+import { setIsAppQuitting, setIsUpdateAvailable } from '../main';
 import { buildContextMenu } from './menu';
 
 class AppUpdater {
