+++ src/utils/safeStorage.ts
